refactor(sandbox-fc): move sandbox tests to external module
